+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Canary health check endpoint
 * Request ID support
 * Additional tests to check underlying Flask application exists and correct configuration used
 * Heroku review applications